Convert to Gitignore.io
Used [Gitignore.io](https://www.gitignore.io/api/java,intellij) to
get a more comprehensive `.gitignore` file that ignores Java and
(some) IntelliJ files. I've left the files that IntelliJ recommends
in to make it easier for students to load the project up in lab and
have the dependencies work.