Add a main that's somewhat useful
This adds a very simple `main()` that prompts the user for a currency code and returns the exchange rate for that code for a fixed date. Nothing fancy, but it works.

I also followed this Stackoverflow solution (https://stackoverflow.com/a/37737186) so that you can run things on the command line through Gradle via:

   ./gradlew run -q --console=plain